Fix quoting problem with lyrics
Setting the lyrics via a "value" attribute in a div doesn't work when
the lyrics text contains quotes. Instead, set non-default lyrics text
as the div child, but if the lyrics text is equal to the default
lyrics, set an empty lyrics text value so that the placeholder value
is displayed instead.

Also, change double quotes to "&quot;" in lyrics HTML.